Remove old centos scripts, start dev of new ones.
All folders are being renamed to be all lowercase, as such the centos folder was changed. The old scripts for centos were all deleted, as they're not needed anymore and a new utilities and install script were added instead. Finally, a shared folders was created to house files common to all OS installs.